// src/app/vis/components/fishing-points-map/layer-field.enum.ts
export enum VHA_WATERCOURSE_FIELD {
    VHAS = 'VHAS',
    VHAG = 'VHAG',
    NAAM = 'NAAM',
    CATC = 'CATC',
    LBLCATC = 'LBLCATC',
    BEKNR = 'BEKNR',
    BEKNAAM = 'BEKNAAM',
    STRMGEB = 'STRMGEB',
    KWALDOEL = 'KWALDOEL',
    LBLKWAL = 'LBLKWAL',
    LBLGEO = 'LBLGEO',
    VHAZONENR = 'VHAZONENR',
    LENGTE = 'LENGTE',
    WTRLICHC = 'WTRLICHC'
}

export enum BRU_WATERCOURSE_FIELD {
    TYPE = 'type',
    NAAM = 'naam',
    BESCHRIJVING = 'beschrijvi',
    CODECAT = 'codecat',
    CATEGORIE = 'categorie'
}

export enum BLUE_LAYER_FIELD {
    WVLC = 'WVLC',
    VERSIE = 'Versie',
    NAAM = 'NAAM',
    WTRLICHC = 'WTRLICHC'
}

export enum TOWN_LAYER_FIELD {
    GEMEENTE = 'GEMEENTE',
    NISCODE = 'NISCODE',
    NISCODE_PR = 'NISCODE_PR',
    PROVINCIE = 'PROVINCIE'
}